+++ src/registry/interfaces/IModuleRegistry.sol
@@ -11,8 +11,8 @@ interface IModuleRegistry {
     event ModuleRegistered(bytes32 indexed key, string name, address factory);
     event ModuleUpdated(bytes32 indexed key, address newFactory);
     event ModuleStatusChanged(bytes32 indexed key, bool active);
-    event OrgModuleSubscribed(uint256 indexed orgId, bytes32 indexed moduleKey);
-    event OrgModuleUnsubscribed(uint256 indexed orgId, bytes32 indexed moduleKey);
+    event OrgModuleSubscribed(uint48 indexed orgId, bytes32 indexed moduleKey);
+    event OrgModuleUnsubscribed(uint48 indexed orgId, bytes32 indexed moduleKey);
 
     // ====== Errors ======
 
@@ -46,24 +46,24 @@ interface IModuleRegistry {
     /**
      * @dev Kích hoạt module cho Organization (khi org trả tiền).
      */
-    function subscribeOrgToModule(uint256 orgId, bytes32 moduleKey) external;
+    function subscribeOrgToModule(uint48 orgId, bytes32 moduleKey) external;
 
     /**
      * @dev Huỷ đăng ký module cho Organization.
      */
-    function unsubscribeOrgFromModule(uint256 orgId, bytes32 moduleKey) external;
+    function unsubscribeOrgFromModule(uint48 orgId, bytes32 moduleKey) external;
 
     // ====== View Functions ======
 
     /**
      * @dev Kiểm tra org đã subscribe module chưa.
      */
-    function isOrgSubscribed(uint256 orgId, bytes32 moduleKey) external view returns (bool);
+    function isOrgSubscribed(uint48 orgId, bytes32 moduleKey) external view returns (bool);
 
     /**
      * @dev Lấy danh sách module keys mà org đã subscribe.
      */
-    function getOrgModules(uint256 orgId) external view returns (bytes32[] memory);
+    function getOrgModules(uint48 orgId) external view returns (bytes32[] memory);
 
     /**
      * @dev Lấy factory address của module.
